backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
anthropic==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.3.0
httpx==0.27.0
langgraph==0.2.28
langchain-core==0.3.10
langchain-anthropic==0.2.3
langchain-community==0.2.16
chromadb==0.5.0
psycopg2-binary==2.9.9
redis==5.0.4
python-multipart==0.0.9
sse-starlette==2.1.0
